[remote_cache] Move resources functions into their respective response classes
